pyproject.toml
[project]
name = "demo-eval-metric"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.2",
    "polars>=1.32.3",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]
